requirements.txt
joblib==1.4.2
numpy==2.2.3
pandas==2.2.3
Pillow==11.1.0
Requests==2.32.3
scikit_learn==1.6.1
scipy==1.15.2
streamlit==1.41.1
